CHORE <codebase refactor> : SINGLE RESPONSIBILITY PRINCIPLE IMPLEMENTATION
This refactor code to reflect single responsibility principle by splitting
functionalities into separate modules.

FINISHED #158948501
