+++ Procfile
@@ -1,1 +1,1 @@
-web: npm run build && npm run migrate && npm start
+web: npm run build && npm start
